prueba ficheros

+++ main.rb
@@ -1,7 +1,15 @@
+$:.unshift "."
 require 'sinatra'
 require "sinatra/reloader" if development?
+require 'sinatra/flash'
+require 'pl0_program'
 require 'data_mapper'
+require 'auth'
 require 'pp'
+
+enable :sessions
+set :session_secret, '*&(^#234)'
+set :reserved_words, %w{grammar test login auth}
 
 # full path!
 DataMapper.setup(:default, 
@@ -45,33 +53,60 @@ end
 post '/save' do
   pp params
   name = params[:fname]
-  c  = PL0Program.first(:name => name)
-  puts "prog <#{c.inspect}>"
-  if c
-    c.source = params["input"]
-    c.save
-  else
-    if PL0Program.count < 10
-        c = PL0Program.new
-        c.name = params["fname"]
+
+  if session[:auth] # authenticated
+    if settings.reserved_words.include? name  # check it on the client side
+      flash[:notice] = 
+        %Q{<div class="error">Can't save file with name '#{name}'.</div>}
+      redirect back
+    else 
+      c  = PL0Program.first(:name => name)
+      if c
         c.source = params["input"]
         c.save
+      else
+        if PL0Program.count < 10
+            c = PL0Program.new
+            c.name = params["fname"]
+            c.source = params["input"]
+            c.save
+        end
       end
+      pp c
+      redirect '/'
+      end
+  else
+    flash[:notice] = 
+      %Q{<div class="error">You are not authenticated.<br />
+         Sign in with Google.
+         </div>}
+    redirect back
   end
-  pp c
-  redirect '/'
 end
-
 
 post '/delete' do
   pp params
   name = params[:fname]
-  c  = PL0Program.first(:name => name)
-  puts "prog <#{c.inspect}>"
-  if c
-    c.source = params["input"]
-    c.destroy
+  
+  if session[:auth] # authenticated
+    if settings.reserved_words.include? name  # check it on the client side
+      flash[:notice] = 
+        %Q{<div class="error">Can't delete file with name '#{name}'.</div>}
+      redirect back
+    else 
+      c  = PL0Program.first(:name => name)
+        if c
+          c.source = params["input"]
+          c.destroy
+        end
+      pp c
+      redirect '/'
+    end
+  else
+    flash[:notice] = 
+      %Q{<div class="error">You are not authenticated.<br />
+         Sign in with Google.
+         </div>}
+    redirect back
   end
-  pp c
-  redirect '/'
 end
